Rewrite of automatic pipelining.
The new implementation uses the same lazy-IO concept as previously.
However the new version:

* Is clearer in what-gets-evaluated-when. Most notably, the spine
  of the list of replies can now be evaluated without forcing the
  replies. The replies are now explicitly forced at the
  appropriate times.
* Does not use an additional evaluation thread.
* Improves pipelining performance by ~5%.
* Get rid of BoundedChan dependency.

One negative is that pipelining is no longer generic in the reply
type. But this would only be necessary in a stand-alone library.

A previous idea was a much more drastic rewrite, using a
Monad/Applicative interface to distinguish between pipelining and
blocking commands, using the ApplicativeDo language extension.
This approach works, but two drawbacks made me decide against it:
* I did not find a way to limit the number of pending replies
  without essentially re-creating the lazy-io version. So
  complexity was not much reduced.
* The user has to enable ApplicativeDo, which would mean
  pipelining does not "just work".