Return list from srandmemberN
`srandmemberN` incorrectly specified its return type as `Maybe
ByteString`, when the point of it is to return a list of items. The
command always returns a multi bulk response even if you specify a
count of 0 or 1 or the key doesn't exist, so there is actually no way
to call this function and not get a `Left`. I have changed that to a
list and added a test to show it works.