Fix xaddOpts
The MAXLEN is supposed to come before the ID. It works in either order
for strict MAXLEN, but for approximate MAXLEN only one order works.

The testXAddRead test case only checks that the stream ends up with 1
element, which doesn't exclude the case where only one of the statements
does anything.