Fix `takeEvery` implementation
The original `takeEvery` implementation wasn't wrong per se. It did
select one out of every `n` elements passed in. It was however
incompatible with the way we are using it. We're passing it a list of
possible keys, the first element of that list definitely being a key. We
need `takeEvery` to select that first element, then every `n`th element
that follows.