Add cluster test boilerplate
Adding this as a separate test endpoint so it's possible to run
non-cluster tests without needing to have a locally running redis
cluster.

We're starting out running the same test suite for clustered mode we do
for non-clustered mode, but that's not going to last. Some Redis
commands we don't currently support in cluster mode, some we might never
be able to support. We'll need to adjust tests to reflect this.