Text fixes for cluster mode
Some tests we cannot run in cluster mode because they use commands we
currently don't support in cluster mode. A lot of tests can be ran if we
ensure the keys touched by the test are in the same hashslot. We can
adopt existing tests by adding a common expression in curly braces to
the keys, which will ensure those keys end up in the same hashslot. This
will not affect how the test runs in non-cluster mode.