Make codegen preserve manual export edits
PR informatikr/hedis#157 manually modified the Commands.hs file by
adding to the export list and slightly changing the import list. When I
tried to regenerate the file, this led to compile errors.

This makes it so that future codegen will preserve these manual edits
and avoid compile issues when the file is regenerated.