Make non-optional list args be NonEmpty lists
This came from an issue with using SREM - SREM will error if the user
tries to pass in an empty list as the collection of values to remove,
even though the type signature allows a list of any size.

To catch this at compile-time, this makes the signature for SREM, and
all other commands with an argument which is `multiple` and not
`optional`, use `Data.List.NonEmpty`.

This forbids some valid uses of commands where an argument is marked
`multiple` but not `optional` - but that needs to be fixed at the
Redis level.