add CLUSTER INFO command and use it for new checkedConnectCluster
Redis nodes in cluster mode can be pinged, but a successful pong from one
cluster node would not determine whether the cluster and thus all its nodes are
up and running. The cluster is up only if the `CLUSTER INFO` command reports
`cluster_state:ok`.

This PR adds the `CLUSTER INFO` command and the `checkedConnectCluster` helper
function for safely connecting to Redis clusters only when the cluster is up.