Fix -Wincomplete-uni-patterns error in stream decoding
Function decodeValue in stream decogin was written using
two functions in a way that it provided temporary value
that had a incomplete match. While it's on two have such
matching because we build only well formed value, but it
can't be proven by the compiler. In this commit we inline
those functions removing a error.